+++ src/afb-evt.h
@@ -28,10 +28,10 @@ struct afb_evt_listener;
 
 struct afb_evt_itf
 {
-	void (*push)(void *closure, const char *event, int evtid, struct json_object *object);
+	void (*push)(void *closure, const char *event, uint16_t evtid, struct json_object *object);
 	void (*broadcast)(void *closure, const char *event, struct json_object *object, const uuid_binary_t uuid, uint8_t hop);
-	void (*add)(void *closure, const char *event, int evtid);
-	void (*remove)(void *closure, const char *event, int evtid);
+	void (*add)(void *closure, const char *event, uint16_t evtid);
+	void (*remove)(void *closure, const char *event, uint16_t evtid);
 };
 
 extern struct afb_evt_listener *afb_evt_listener_create(const struct afb_evt_itf *itf, void *closure);
@@ -50,7 +50,7 @@ extern struct afb_evtid *afb_evt_evtid_addref(struct afb_evtid *evtid);
 extern void afb_evt_evtid_unref(struct afb_evtid *evtid);
 
 extern const char *afb_evt_evtid_fullname(struct afb_evtid *evtid);
-extern int afb_evt_evtid_id(struct afb_evtid *evtid);
+extern uint16_t afb_evt_evtid_id(struct afb_evtid *evtid);
 
 extern const char *afb_evt_evtid_name(struct afb_evtid *evtid);
 
@@ -62,11 +62,10 @@ extern int afb_evt_watch_add_evtid(struct afb_evt_listener *listener, struct afb
 extern int afb_evt_watch_sub_evtid(struct afb_evt_listener *listener, struct afb_evtid *evtid);
 
 
-
 extern struct afb_event_x2 *afb_evt_event_x2_create(const char *fullname);
 extern struct afb_event_x2 *afb_evt_event_x2_create2(const char *prefix, const char *name);
 extern const char *afb_evt_event_x2_fullname(struct afb_event_x2 *event);
-extern int afb_evt_event_x2_id(struct afb_event_x2 *eventid);
+extern uint16_t afb_evt_event_x2_id(struct afb_event_x2 *eventid);
 extern struct afb_event_x2 *afb_evt_event_x2_addref(struct afb_event_x2 *eventid);
 extern void afb_evt_event_x2_unref(struct afb_event_x2 *eventid);
